requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
certifi==2026.4.22
frozenlist==1.8.0
idna==3.13
magic-filter==1.0.12
multidict==6.7.1
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.2.2
typing_extensions==4.15.0
yarl==1.23.0
